Convert readwrite trigger related code to condition monitor
 - Cut and paste methods related to monitoring/triggering a task from ReadWrite
 file to ConditionMonitor file
 - Rename readwrite_test.go to condition_monitor_test.go (all remaining tests
 are about monitoring/triggering a task)
